Avoid duplicates in lockers
In cases where locker is holding several locks on the same entry,
make sure to use only the highest lock to avoid duplicates.

Lock mode is transfered into numeric form via array. It is possible
to convert it back, but as mode is not shown in the final output
I choose not bother.