Another duplicates fix for locks tree
It turns out that quite frequently single session will block others on
different objects, especially if it is idle in transaction. Therefore
get rid of the `mode` column in `leads` and `tree` subqueries and group
leads only by locker.

This lead to the change for the non-cyclic sessions. As `mode` is gone,
lockers are filtered by highest level instead.

Last change was to get rid of early cycle checks in the `leads` subquery,
as it is causing wrong results when tree is formed.