Provide `wait_on_object` column in the output
It might desirable to see which object we're waiting on.
To do so join to the pairs by both, locker and waiter, otherwise
duplicates are prodcued.
To do so add `dad` column to the `tree` subquery.

It was tempting to also change database and relation (wherever possible)
oids into names, but I choose not to do so.